meanvc_gui/requirements.txt
# MeanVC Desktop App (PySide6)
# Install with: pip install -r requirements.txt

# GUI
PySide6>=6.5.0
PySide6-Addons>=6.5.0
PySide6-Charts>=6.5.0

# Audio processing
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
sounddevice>=0.4.0
Pillow>=10.0.0

# MeanVC dependencies (from parent project)
librosa>=0.10.0
tqdm>=4.65.0
soundfile>=0.12.0

# Development
pytest>=7.0.0
